Move test classes to the same package as the SUT
